pyproject.toml
[project]
name = "quant101"
version = "0.1.0"
description = ""
authors = [
    {name = "JerryHong",email = "huntcamilla06@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "boto3 (>=1.40.7,<2.0.0)",
    "botocore (>=1.40.7,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "polars (>=1.32.2,<2.0.0)",
    "pyarrow (>=21.0.0,<22.0.0)",
    "duckdb (>=1.3.2,<2.0.0)",
    "polygon-api-client (>=1.15.3,<2.0.0)",
    "yfinance (>=0.2.65,<0.3.0)",
    "seolpyo-mplchart (>=1.4.1.3,<2.0.0.0)",
    "ipywidgets (>=8.1.7,<9.0.0)",
    "longport (>=3.0.12,<4.0.0)",
    "exchange-calendars (>=4.11.1,<5.0.0)",
    "ta-lib (>=0.6.6,<0.7.0)",
    "s3fs (>=2025.9.0,<2026.0.0)",
    "pytorch-tabnet (>=4.1.0,<5.0.0)",
    "lightgbm (>=4.6.0,<5.0.0)",
    "ta (>=0.11.0,<0.12.0)",
    "optuna (>=4.5.0,<5.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "tensorflow (>=2.20.0,<3.0.0)",
]

[tool.poetry]
packages = [{include = "*", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pre-commit = "^4.3.0"
ipykernel = "^6.30.1"
black = "^25.1.0"
isort = "^6.0.1"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--ignore=data --ignore=scripts/run_backtest.py"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
